Metadata addition working for Composite models
