+++ examples/phi3_5/qdq_config_vitis_ai.json
@@ -53,10 +53,17 @@
         },
         "sp": { "type": "SplitModel" },
         "st": { "type": "StaticLLM", "batch_size": 1, "context_length": 64 },
-        "p": { "type": "VitisAIAddProviderGenAIConfig", "provider_options": { "log_level": "info" } }
+        "p": { "type": "VitisAIUpdateGenAIProviderOpts", "provider_options": { "log_level": "info" } },
+        "addmetadata": {
+            "type": "VitisAIAddMetaData",
+            "config_meta_data_keys": [ "architectures", "model_type" ],
+            "activation_type": "QUInt16",
+            "weight_type": "QUInt8",
+            "quant_type": "OnnxStaticQuantization"
+        }
     },
     "log_severity_level": 1,
     "output_dir": "models/phi3_5_qdq",
-    "cache_dir": "phi3_5_cache",
+    "cache_dir": "cache/phi3_5_cache",
     "no_artifacts": true
 }
